apps/backend/pyproject.toml
[tool.poetry]
name = "lab-backend"
version = "0.2.0"
description = "FastAPI backend for laboratory works (Lab2)"
authors = ["Lab <lab@example.local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pydantic = "^2.7.0"
loguru = "^0.7.2"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pydantic-settings = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


